requirements.txt
cytoolz==0.11.0
flatbuffers==1.12
fonttools==4.25.0
keras==2.9.0
Keras-Preprocessing==1.1.2
matplotlib==3.5.1
statsmodels==0.13.2
tensorflow-cpu==2.9.1
voxelmorph==0.2
xgboost==0.90
Flask==2.0.2
Jinja2==3.0.2
click==7.1.2
joblib==0.17.0
gunicorn==20.1.0
opencv-python==4.6.0.66
lime==0.2.0.1